* [test] update test
