+++ test/run-android-ci.sh
@@ -1,4 +1,5 @@
 ./test/serve.sh &
 npm install macaca-cli --save-dev
 npm install macaca-android --save-dev
+macaca doctor
 ./test/run.sh
